requirements.txt
# ==============================================================
# Requirements — pinned to known-good versions for production.
# ==============================================================

# ── Core API ────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart>=0.0.20

# ── ML / CLIP ───────────────────────────────────────────────
torch>=2.1.0
open-clip-torch==2.29.0
Pillow>=10.0.0

# ── Vector DB ───────────────────────────────────────────────
qdrant-client==1.13.2

# ── Cache ───────────────────────────────────────────────────
redis>=5.0.0

# ── ONNX Runtime (optional, set USE_ONNX=true to enable) ──
onnx==1.17.0
onnxruntime==1.19.2

# ── Observability ──────────────────────────────────────
structlog>=24.0.0

# ── OpenTelemetry ─────────────────────────────────────
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp-proto-grpc>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-logging>=0.46b0

# ── Auth / Rate Limiting ──────────────────────────────
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9

# ── Serialization ─────────────────────────────────────
orjson>=3.9.0

# ── Utilities ──────────────────────────────────────────
numpy>=1.24.0
tqdm>=4.66.0
psutil>=5.9.0
python-dotenv>=1.0.0

# ── LLM (Cerebras + Groq via OpenAI-compatible SDK) ───────
openai>=1.0.0

# ── PDF Processing ────────────────────────────────────────
PyMuPDF>=1.24.0

# ── Text Embeddings ───────────────────────────────────────
sentence-transformers>=3.0.0

# ── SSE Streaming ─────────────────────────────────────────
sse-starlette>=2.0.0

# ── V2: Unified Embeddings (Jina-CLIP v2) ─────────────────
transformers>=4.40.0
einops>=0.8.0

# ── V2: VLM (SmolVLM-500M) ────────────────────────────────
# Uses transformers + torch (already listed above)
aiofiles>=24.0.0

# ── V2: Knowledge Graph ───────────────────────────────────
networkx>=3.1

# ── V2: Web Scraping ──────────────────────────────────────
langdetect>=1.0.9
youtube-transcript-api>=0.6.0
firecrawl-py>=0.1.0

# ── V2: Cross-Encoder Reranker ─────────────────────────────
# Uses sentence-transformers CrossEncoder (already listed above)

# ── Dev / Testing ─────────────────────────────────────────
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
